Cleanup + Fixes + Version Bump for later

+++ java/org/silvercatcher/reforged/items/weapons/ItemKeris.java
@@ -44,7 +44,7 @@ public class ItemKeris extends ItemSword implements ItemExtension {
 		            while (amount > 0) {
 		                int j = EntityXPOrb.getXPSplit(amount);
 		                amount -= j;
-		                w.spawnEntityInWorld(new EntityXPOrb(w, (double)e.posX + 0.5D, (double)e.posY + 0.5D, (double)e.posZ + 0.5D, j));
+		                w.spawnEntityInWorld(new EntityXPOrb(w, e.posX + 0.5D, e.posY + 0.5D, e.posZ + 0.5D, j));
 		            }
 		        }
 			}
